# swagger.yaml
openapi: 3.0.3
info:
  title: EduPlanner API
  description: API para la plataforma educativa EduPlanner - Planeaciones didácticas con IA
  version: 1.0.0
  contact:
    name: EduPlanner Support
    email: support@eduplanner.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo
  - url: https://eduplanner.vercel.app/api
    description: Servidor de producción
paths:
...
  /messages:
    get:
      tags:
        - Mensajes
      summary: Obtener mensajes del usuario
      description: Recupera todos los mensajes asociados a un usuario específico
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
          description: ID del usuario
      responses:
        '200':
          description: Lista de mensajes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    user_id:
                      type: string
                    content:
                      type: string
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        '400':
          description: User ID es requerido
        '500':
          description: Error del servidor
  /save-message:
    post:
      tags:
        - Mensajes
      summary: Guardar mensaje
      description: Guarda un nuevo mensaje en la base de datos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                content:
                  type: string
                type:
                  type: string
              required:
                - user_id
                - content
      responses:
        '200':
          description: Mensaje guardado exitosamente
        '400':
          description: Datos requeridos faltantes
        '500':
          description: Error del servidor
...
  /quality-feedback:
    post:
      tags:
        - Feedback
      summary: Feedback de calidad
      description: Endpoint para retroalimentación específica sobre la calidad del contenido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content_id:
                  type: string
                quality_score:
                  type: integer
                  minimum: 1
                  maximum: 10
                comments:
                  type: string
      responses:
        '200':
          description: Feedback de calidad registrado
        '500':
          description: Error del servidor
...
  /stripe/webhook:
    post:
      tags:
        - Pagos
      summary: Webhook de Stripe
      description: Endpoint para recibir eventos de Stripe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook procesado exitosamente
        '400':
          description: Webhook inválido
  /stripe/cancel-subscription:
    post:
      tags:
        - Pagos
      summary: Cancelar suscripción
      description: Cancela la suscripción activa del usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription_id:
                  type: string
              required:
                - subscription_id
      responses:
        '200':
          description: Suscripción cancelada exitosamente
        '400':
          description: ID de suscripción requerido
        '500':
          description: Error del servidor
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticación JWT de Supabase
security:
  - BearerAuth: []
...